scripts/revdep-paxmark: tool to find mismatched pax markings
This tool will search for binaries which have different pax
markings than the libraries they link against.  Since running
processes run with the pax enforcings of the binary, not the
libraries they link against, this utility can be used to locate
such binaries and mark them with the same flags as their libraries.

This code is derived from revdep-rebuild.