src/paxctl-ng.c: do not use '-' when setting null XATTR PAX flags
There was a bug in the implementation of XATTR PAX flags where it
was assumed that '-' would be ignored in kernel.  However, anything
other than PpEeMmRrSs is considered invalid and results in a fail.
See pax_parse_xattr_pax() in fs/binfmt_elf.c for a pax patched kernel.

The pxtpax test was updated as well.

X-Gentoo-Bug: 446518
X-Gentoo-Bug-URL: https://bugs.gentoo.org/446518