misc/install-xattr: refine the behavior of which()
The wrapper must avoid invoking itself in its attempt to find the
system `install` path.  The old behavior was to avoid looking in
the same directory as the wrapper, but this is too crude since we
might have the wrapper in /usr/bin (say), and the system `install`
in the same directory, in which case we'll wind up skipping it.
The new behavior of which() is to accept argv[0], find its canonical
path and skip if it is equal to the canditate system `install` we
are checking.

We also add PORTAGE_BIN_PATH which will be used in a future commit
to avoid this wrapper from finding portage's wrapper.  We can then
get into a situation where the portage wrapper, usually at
${PORTAGE_BIN_PATH}/ebuild-helpers/xattr/install calls this wrapper,
usually at /usr/bin/install-xattr, which in turn calls the portage
wrapper in an infinite self-invocation (fork bomb).