Correctly determine dst path if src is in a dir
There is a mismatch between how install and install-xattr determines
the destination path.

When running `install src/a dst/' the correct output is `dst/a'.

install-xattr was incorrectly joining the paths so thought the output
was `dst/src/a' which did not exist.

A test case has been added to the test script too.

Signed-off-by: Jason Zaman <jason@perfinion.com>
Signed-off-by: Anthony G. Basile <blueness@gentoo.org>