misc/install-xattr: correct potential fork bomb
The which() function compares portage_helper_path, to canpath
and skips it when appropriate:

	if (portage_helper_path)
		if (!strcmp(portage_helper_path, canpath))
			goto skip;

However, portage_helper_path has not been canonicalized with
the realpath function, so strcmp can return false even though
the paths are equivalent.  This may occurs when /usr/lib is a
symlink to /usr/lib64.

X-Gentoo-Bug: 523994
X-Gentoo-Bug-URL: https://bugs.gentoo.org/523994